Add to MockTabWebView
